refactor(phase-4): client extraction of handlers (#100)
## Problem

`Anubis.Client.Base` module is currently massive, this is a first try to
reduce it

## Solution

Extracted sampling capability and other message handlers (tool calls
responses for ex) to separate modules

## Rationale

N/A
