+++ .github/workflows/release.yml
@@ -39,7 +39,7 @@ jobs:
       - name: Install Zig
         uses: goto-bus-stop/setup-zig@v2
         with:
-          version: 0.14.1
+          version: 0.15.2
 
       - name: Install XZ
         run: sudo apt-get update && sudo apt-get install -y xz-utils
